Add USB Serial
Made the USB UART example work with the rest of
the PIOs acting as UARTs too. It seems that unless
the `usb_dev.poll()` is called, the board is not
picked up by `lsusb`, nor is it's driver loaded.
Not a problem anyway, since the USB UART will be
polled anyway.